added require property, handled permission errors and added ticket button handler

+++ src/commands/SlashCommands/tickets/ticketCommands.ts
@@ -7,6 +7,7 @@ import { ticketModel } from "../../../models/ticket";
 export default new slashCommand({
     name: "ticket",
   description: "allows to modify ticket config / execute ticket commands",
+  require:["mongooseConnectionString"],
   userPermissions: ["MANAGE_THREADS"],
   botPermissions: ["MANAGE_THREADS"],
   options: [
@@ -208,11 +209,11 @@ export default new slashCommand({
             name: config,
             guildId: guildId,
           });
-
           if (existing.length == 0) return interaction.followUp({content: `ticket does not exist`});
           ticketModel.deleteOne({guildId: guildId, name: config})
             .then(() => {
               const message = (interaction.guild.channels.cache.get(existing[0].channel) as GuildTextBasedChannel).messages.cache.get(existing[0].linkedmessage)
+              if (!message) return interaction.followUp({content: `I could not delete the linked message, please do it yourself`})
               message.delete().then(() => interaction.followUp({content: `successfully deleted`}))})
         }
       }
